restore postcss as we need exact position of urls to use __v__
