+++ packages/ast/src/js/js_static_analysis/web_worker.js
@@ -22,10 +22,10 @@ export const analyzeImportScriptCalls = (node, { onUrl }) => {
         subtype: "self_import_scripts_arg",
         expectedType: "js_classic",
         specifier: specifierNode.value,
-        specifierStart: specifierNode.start,
-        specifierEnd: specifierNode.end,
-        specifierLine: specifierNode.loc.start.line,
-        specifierColumn: specifierNode.loc.start.column,
+        start: specifierNode.start,
+        end: specifierNode.end,
+        line: specifierNode.loc.start.line,
+        column: specifierNode.loc.start.column,
       })
     }
   })
