+++ tests/dev_and_build/import_assert_type_css/import_assert_type_css_build.test.mjs
@@ -5,16 +5,20 @@ import { jsenvPluginMinification } from "@jsenv/plugin-minification"
 import { build } from "@jsenv/core"
 import { startFileServer } from "@jsenv/core/tests/start_file_server.js"
 import { executeInChromium } from "@jsenv/core/tests/execute_in_chromium.js"
+import {
+  readSnapshotsFromDirectory,
+  writeSnapshotsIntoDirectory,
+} from "@jsenv/core/tests/snapshots_directory.js"
 
-const test = async (options) => {
-  const { buildManifest } = await build({
+const test = async ({ snapshotsDirectoryUrl, ...rest }) => {
+  const { buildFileContents, buildManifest } = await build({
     logLevel: "warn",
     rootDirectoryUrl: new URL("./client/", import.meta.url),
     buildDirectoryUrl: new URL("./dist/", import.meta.url),
     entryPoints: {
       "./main.html": "main.html",
     },
-    ...options,
+    ...rest,
   })
   const server = await startFileServer({
     rootDirectoryUrl: new URL("./dist/", import.meta.url),
@@ -22,53 +26,57 @@ const test = async (options) => {
   const { returnValue } = await executeInChromium({
     url: `${server.origin}/main.html`,
     /* eslint-disable no-undef */
-    pageFunction: async () => {
-      return window.resultPromise
-    },
+    pageFunction: async () => window.resultPromise,
     /* eslint-enable no-undef */
   })
-  const actual = returnValue
+  const snapshotsFileContent = readSnapshotsFromDirectory(snapshotsDirectoryUrl)
+  writeSnapshotsIntoDirectory(snapshotsDirectoryUrl, buildFileContents)
+  const actual = {
+    snapshotsFileContent,
+    returnValue,
+  }
   const expected = {
-    bodyBackgroundColor: "rgb(255, 0, 0)",
-    bodyBackgroundImage: `url("${server.origin}/${buildManifest["other/jsenv.png"]}")`,
+    snapshotsFileContent: buildFileContents,
+    returnValue: {
+      bodyBackgroundColor: "rgb(255, 0, 0)",
+      bodyBackgroundImage: `url("${server.origin}/${buildManifest["other/jsenv.png"]}")`,
+    },
   }
   assert({ actual, expected })
 }
 
 // can use <script type="module">
 await test({
+  snapshotsDirectoryUrl: new URL("./snapshots/js_module/", import.meta.url),
   runtimeCompat: { chrome: "89" },
   plugins: [jsenvPluginBundling()],
 })
-// no bundling
-await test({ runtimeCompat: { chrome: "65" } })
-// cannot use <script type="module">
-await test({
-  runtimeCompat: { chrome: "60" },
-  plugins: [jsenvPluginBundling()],
-})
-
-// minification
-{
-  const { buildInlineContents } = await build({
-    logLevel: "warn",
-    rootDirectoryUrl: new URL("./client/", import.meta.url),
-    buildDirectoryUrl: new URL("./dist/", import.meta.url),
-    entryPoints: {
-      "./main.html": "main.html",
-    },
-    plugins: [
-      jsenvPluginBundling(),
-      jsenvPluginMinification({
-        js_module: false,
-        css: true,
-      }),
-    ],
-  })
-  const cssKey = Object.keys(buildInlineContents).find((key) =>
-    key.endsWith(".css"),
-  )
-  const actual = buildInlineContents[cssKey]
-  const expected = `body{background-color:red;background-image:url("+__v__("/other/jsenv.png")+")}`
-  assert({ actual, expected })
-}
+// can use <script type="module"> + no bundling
+// await test({
+//   snapshotsDirectoryUrl: new URL(
+//     "./snapshots/js_module_no_bundling/",
+//     import.meta.url,
+//   ),
+//   runtimeCompat: { chrome: "65" },
+// })
+// // can use <script type="module"> + minification
+// await test({
+//   snapshotsDirectoryUrl: new URL(
+//     "./snapshots/js_module_css_minified/",
+//     import.meta.url,
+//   ),
+//   runtimeCompat: { chrome: "60" },
+//   plugins: [
+//     jsenvPluginBundling(),
+//     jsenvPluginMinification({
+//       js_module: false,
+//       css: true,
+//     }),
+//   ],
+// })
+// // cannot use <script type="module">
+// await test({
+//   snapshotsDirectoryUrl: new URL("./snapshots/systemjs/", import.meta.url),
+//   runtimeCompat: { chrome: "60" },
+//   plugins: [jsenvPluginBundling()],
+// })
