+++ tests/dev_and_build/import_assert_type_css/snapshots/js_module/main.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html><html><head>
+    <title>Title</title>
+    <meta charset="utf-8">
+    <link rel="icon" href="data:,">
+    <script>
+;(function() {
+  var __versionMappings__ = {
+  "/other/jsenv.png": "/other/jsenv.png?v=25e95a00"
+};
+  window.__v__ = function (specifier) {
+    return __versionMappings__[specifier] || specifier
+  };
+})();
+</script>
+  </head>
+
+  <body>
+    <script no-html-supervisor="">
+      window.resultPromise = new Promise((resolve) => {
+        window.resolveResultPromise = resolve
+      })
+    </script>
+    <script type="module" src="/js/main.js?v=cfdedff1"></script>
+  
+
+</body></html>
